+++ core/src/main/java/org/apache/carbondata/core/indexstore/row/DataMapRow.java
@@ -19,7 +19,8 @@ package org.apache.carbondata.core.indexstore.row;
 import org.apache.carbondata.core.indexstore.schema.DataMapSchema;
 
 /**
- * Index row
+ * It is just a normal row to store data. Implementation classes could be safe and unsafe.
+ * TODO move this class a global row and use across loading after DataType is changed class
  */
 public abstract class DataMapRow {
